requirements.txt
# PIVO Dependencies

# Google Gemini API
google-generativeai>=0.8.0

# Hive connectivity (without native SASL - using NOSASL mode)
pyhive>=0.7.0
thrift>=0.16.0

# HDFS access
hdfs>=2.7.0

# Spark
pyspark>=3.5.0

# Kafka (optional, for ingestion listener)
kafka-python>=2.0.2

# Utilities
python-dotenv>=1.0.0
